Bug #17 fixed; changes in hmmsearch.c, PostprocessSignificantHit returns sc. bug17.sh exercise added. --cpu0 dump bug in hmmsearch fixed; pvm/thread conditional code handled by variables instead of ifdef, to simplify logic. some -Wall warnings silenced. ./configure --enable-debugging without arg activates -g -Wall cflags xref STL6 p.1
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/trunk/hmmer@728 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/display.c
@@ -293,7 +293,7 @@ DisplayPlan7Matrix(char *dsq, int L, struct plan7_s *hmm, struct dpmatrix_s *mx)
   int k;
 
   printf("         *      ");
-  for (i=1;i<=L;i++) printf("    %c      ",Alphabet[dsq[i]]);
+  for (i=1;i<=L;i++) printf("    %c      ",Alphabet[(int)dsq[i]]);
   printf("\nN    ");
   for (i=0;i<=L;i++) PrintIscore(mx->xmx[i][XMN]);
   for (k=1;k<=hmm->M;k++) {
